refactor(link): split monolithic service into focused modules
Extract link.service.ts (1400+ lines) into smaller, focused modules:
- link-crud.service.ts: CRUD operations
- link-analytics.service.ts: Analytics and statistics
- link-bulk.service.ts: Bulk operations
- link-password.service.ts: Password management
- link-status.service.ts: Status toggles
- link-utils.service.ts: Export and utility functions
- link-shared.ts: Shared utilities
- index.ts: Barrel file for backward-compatible exports
